requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
ecdsa==0.19.2
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.10
pwdlib==0.3.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
